pyproject.toml
[project]
name = "wheelspin-gif-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "matplotlib>=3.7.5",
    "numpy>=1.24.4",
    "pillow>=10.4.0",
]
